--- tests/tools/class_registry/__init__.py
@@ -1,1 +1,0 @@
-
